Revert "cros_bundle_firmware: Add a 'firmware-type' in each firmware's fdt"
This reverts commit 85aba960bcd75638ebcafbef8e6c4f765f4ad054.

BUG=chrome-os-partner:14545
BRANCH=link
TEST=none

The original commit made an artificial distinction between the two RW BIOS
images to help with debugging. We need to take that back out so that they're
identical again. There *shouldn't* be any difference, but if the bits are
different there's a nagging suspicion that they just *might* behave
differently.

Change-Id: Id6bdf3e62ed50bb0bd7a8bf2f0b2a9711d079d53
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/36511
Reviewed-by: Simon Glass <sjg@chromium.org>