cros_write_firmware: Accept single or double quotes around console output.
When getting the console output using dut-control, the output may be
surrounded by single or double quotes, depending on what's actually in the
output. It seems to default to ', but if there are any 's in the output it
will use ". It's not clear what will happen if there are both 's and "s.

BUG=chrome-os-partner:22103
TEST=With this change, console output is recongized correctly. Before this
change, console output surrounded by ""s was not.
BRANCH=pit,spring

Change-Id: Ia9b7f57adab736d90548d20bc2a728b539a66dfc
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/66793
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>