cros_bundle_firmware: Make storm coreboot image size hack the default
This patch generalizes the method of using the coreboot (CBFS) image
size defined in the fmap.dts file (instead of hardcoding it to the last
MB of the file) that was introduced for Storm in commit 9f36e714 (cbf:
accommodate storm coreboot layout) and later extended to Rush-based
boards. This method is certainly more correct/extensible than the
previous one and a better stopgap until the eventual complete overhaul
of cros_bundle_firmware. It also allows coreboot to pad its output
coreboot.rom file to the actual (physical) ROM size on non-x86 boards.

All fmap.dts definitions for existing boards (except for the three that
were already using this method before) specify their ro-boot section
correctly as either the first or the last MB of the image (and in the
former case produce a 1MB coreboot.rom), so they will pick up the exact
same bytes with this method. The only exception is Cosmos with 0.5MB at
the beginning of the image, which should still produce the same results
since the old coreboot.rom image for Cosmos was also only 0.5MB in size.

BUG=None
TEST=Booted on Veyron_Pinky, Nyan_Blaze and Falco. Compiled Cosmos
and noticed that it doesn't build on ToT anyway.

Change-Id: Ic957150a7e93db1bad6be1a8af2d71d6b68a02f0
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/231475
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/236112
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>
Tested-by: Vadim Bendebury <vbendeb@chromium.org>