cbf: do not assume that u-boot is always there
The cros_bundle_firmware tool once started as a utility to pack u-boot
with the rest of Chrome OS bootprom components. Nowadays u-boot is not
essential in the image, but the utility still expects it to be
present. This requires the user to jump through some hoops when
bundling a coreboot/depthcharge based image.

Let's not try using coreboot if it is not supplied in the command
line, this will make the use of this utility much simpler and help to
accommodate multiple CBFS sections in the same bootporm later.

This also fixes a bug when gbb blob was originally the u-boot as the
source file (it was overwritten later in the code, the second
assignment is being removed in this patch).

BRANCH=storm
BUG=chrome-os-partner:34161
TEST='emerge-storm chromeos-bootimage' still produces a bootable
     Chrome OS firmware.

Change-Id: Id985a0f1d23642191066b7a400cf17ec589f20df
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/237223
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/237950