bundle_firmware: limit size of CBFS for RW slots
Now that boards are switched over to a CBFS in each RW
slot the hashing taking place is a lot more than the
component method. The reason is that 1. there's more
things within the CBFS 2. the RW slots weren't always
fully utilized. To gain back more boot speed limit
the size of the CBFS that is hashed in the RW slots.

This is accomplished by taking advatage of the cbfstool
'compact' command which removes empty file fragmentation
from CBFS regions. This command is invoked for nodes
which contin a cbfs-files subnode. Additionally, the
cbfstool 'print' command is ran with a parseable option
which allows limiting the CBFS region to be hashed by
changing the size of the region to the offset of the
last file if it is empty. i.e. if the last file entry
is empty there's no need to include it in the hashed
data.

Lastly, there's no reason to include a master header
within the RW slots so remove that as well.

BUG=chrome-os-partner:49764
BUG=chromium:445938
BRANCH=glados
CQ-DEPEND=CL:324094,CL:324095
TEST=Built and booted chell with cbfstool and dev-util patches.

Change-Id: I1be6b3c7f3a780bff343e32333ba790f42ad09b8
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Previous-Reviewed-on: https://chromium-review.googlesource.com/324023
(cherry picked from commit 152e76150ee0ebadf71716be8afa0efeae999e3f)
Reviewed-on: https://chromium-review.googlesource.com/325055
Tested-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Commit-Queue: Duncan Laurie <dlaurie@chromium.org>